+++ src/annotations/markups/bidirectional/mathUtils.js
