+++ src/annotations/markups/bidirectional/moveBidirectionalHandles.js
