+++ src/annotations/markups/bidirectional/moveLongAxisStartHandle.js
@@ -1,5 +1,7 @@
 import { distance, intersectLine } from "./mathUtils";
 
+let lastValidCoords;
+
 export default function (handle, shortAxisGeometry, longAxisGeometry) {
   const shortAxisCoords = shortAxisGeometry.getCoordinates();
   const longAxisCoords = longAxisGeometry.getCoordinates();
@@ -25,34 +27,38 @@ export default function (handle, shortAxisGeometry, longAxisGeometry) {
     },
   };
 
-
   const intersection = intersectLine(longAxis, shortAxis);
+  if (!intersection) {
+    longAxisGeometry.setCoordinates(lastValidCoords);
+  }
 
-  const distanceFromShortAxisP1 = distance(shortAxis.start, intersection);
-  const distanceFromShortAxisP2 = distance(shortAxis.end, intersection);
-
-  const distanceToLineP2 = distance(longAxis.start, intersection);
-  const newLineLength = distance(longAxis.start, handle);
+  const distanceToLineP2 = distance(longAxis.end, intersection);
+  const newLineLength = distance(longAxis.end, handle);
 
   if (newLineLength <= distanceToLineP2) {
     return false;
   }
 
-  const dx = (longAxis.start.x - handle.x) / newLineLength;
-  const dy = (longAxis.start.y - handle.y) / newLineLength;
+  lastValidCoords = longAxisCoords;
+
+  const distanceFromShortAxisP1 = distance(shortAxis.start, intersection);
+  const distanceFromShortAxisP2 = distance(shortAxis.end, intersection);
+
+  const dx = (longAxis.end.x - handle.x) / newLineLength;
+  const dy = (longAxis.end.y - handle.y) / newLineLength;
 
   const k = distanceToLineP2 / newLineLength;
 
   const newIntersection = {
-    x: longAxis.start.x + (handle.x - longAxis.start.x) * k,
-    y: longAxis.start.y + (handle.y - longAxis.start.y) * k,
+    x: longAxis.end.x + (handle.x - longAxis.end.x) * k,
+    y: longAxis.end.y + (handle.y - longAxis.end.y) * k,
   };
 
-  shortAxis.start.x = newIntersection.x + distanceFromShortAxisP1 * dy;
-  shortAxis.start.y = newIntersection.y - distanceFromShortAxisP1 * dx;
+  shortAxis.start.x = newIntersection.x - distanceFromShortAxisP1 * dy;
+  shortAxis.start.y = newIntersection.y + distanceFromShortAxisP1 * dx;
 
-  shortAxis.end.x = newIntersection.x - distanceFromShortAxisP2 * dy;
-  shortAxis.end.y = newIntersection.y + distanceFromShortAxisP2 * dx;
+  shortAxis.end.x = newIntersection.x + distanceFromShortAxisP2 * dy;
+  shortAxis.end.y = newIntersection.y - distanceFromShortAxisP2 * dx;
 
   shortAxisGeometry.setCoordinates([
     [shortAxis.start.x, shortAxis.start.y],
