+++ src/annotations/markups/bidirectional/moveShortAxisEndHandle.js
