+++ src/annotations/markups/bidirectional/moveShortAxisStartHandle.js
